+++ src/FrozenWasteland.cpp
@@ -26,8 +26,8 @@ void init(rack::Plugin *p) {
 	p->addModel(modelPhasedLockedLoop);
 	p->addModel(modelPortlandWeather);
 	p->addModel(modelQuadAlgorithmicRhythm);
-	p->addModel(modelQuadGrooveExpander);
-	p->addModel(modelQuadRhythmExpander);
+	p->addModel(modelQARGrooveExpander);
+	p->addModel(modelQARProbabilityExpander);
 	p->addModel(modelQuantussyCell);
 	p->addModel(modelRouletteLFO);
 	p->addModel(modelSeriouslySlowLFO);